Move all classes to their own files
