+++ gTile@shuairan/files/gTile@shuairan/3.8/gTile.js
@@ -40,7 +40,6 @@ __webpack_require__.r(__webpack_exports__);
 
 // EXPORTS
 __webpack_require__.d(__webpack_exports__, {
-  "Grid": () => (/* binding */ Grid),
   "area": () => (/* binding */ extension_area),
   "disable": () => (/* binding */ disable),
   "enable": () => (/* binding */ enable),
@@ -49,7 +48,9 @@ __webpack_require__.d(__webpack_exports__, {
   "getMonitorKey": () => (/* binding */ getMonitorKey),
   "getNotFocusedWindowsOfMonitor": () => (/* binding */ getNotFocusedWindowsOfMonitor),
   "getUsableScreenArea": () => (/* binding */ getUsableScreenArea),
+  "gridSettingsButton": () => (/* binding */ gridSettingsButton),
   "grids": () => (/* binding */ grids),
+  "hideTiling": () => (/* binding */ hideTiling),
   "init": () => (/* binding */ init),
   "move_maximize_window": () => (/* binding */ move_maximize_window),
   "move_resize_window": () => (/* binding */ move_resize_window),
@@ -592,8 +593,8 @@ class TopBar {
     }
 }
 
-;// CONCATENATED MODULE: ./src/3_8/extension.ts
-var extension_decorate = (undefined && undefined.__decorate) || function (decorators, target, key, desc) {
+;// CONCATENATED MODULE: ./src/3_8/ui/Grid.ts
+var Grid_decorate = (undefined && undefined.__decorate) || function (decorators, target, key, desc) {
     var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
     if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
@@ -609,11 +610,347 @@ var extension_decorate = (undefined && undefined.__decorate) || function (decora
 
 
 
+
+const Grid_St = imports.gi.St;
+const Grid_Main = imports.ui.main;
+const Grid_Tweener = imports.ui.tweener;
+const Clutter = imports.gi.Clutter;
+let toggleSettingListener;
+let Grid = class Grid {
+    constructor(monitor_idx, monitor, title, cols, rows) {
+        this.tableWidth = 220;
+        this.tableHeight = 200;
+        this.borderwidth = 2;
+        this.bindFns = {};
+        this.rowKey = -1;
+        this.colKey = -1;
+        this.isEntered = false;
+        this.interceptHide = false;
+        this._initGridSettingsButtons = () => {
+            this.bottombar.destroy_children();
+            let rowNum = 0;
+            let colNum = 0;
+            for (var index = 0; index < gridSettingsButton.length; index++) {
+                if (colNum >= 4) {
+                    colNum = 0;
+                    rowNum += 2;
+                }
+                let button = gridSettingsButton[index];
+                button = new GridSettingsButton(button.text, button.cols, button.rows);
+                this.bottombar.add(button.actor, { row: rowNum, col: colNum, x_fill: false, y_fill: false });
+                button.actor.connect('notify::hover', this._onSettingsButton);
+                colNum++;
+            }
+        };
+        this._displayElements = () => {
+            this.elements = [];
+            let width = this.tableWidth / this.cols - 2 * this.borderwidth;
+            let height = this.tableHeight / this.rows - 2 * this.borderwidth;
+            this.elementsDelegate = new GridElementDelegate();
+            this.elementsDelegate.connect('resize-done', this._onResize);
+            for (let r = 0; r < this.rows; r++) {
+                for (let c = 0; c < this.cols; c++) {
+                    if (c === 0) {
+                        this.elements[r] = [];
+                    }
+                    let element = new GridElement(this.monitor, width, height, c, r, this.elementsDelegate);
+                    this.elements[r][c] = element;
+                    this.table.add(element.actor, { row: r, col: c, x_fill: false, y_fill: false });
+                    element.show();
+                }
+            }
+        };
+        this.refresh = () => {
+            this.table.destroy_all_children();
+            this.cols = preferences.nbCols;
+            this.rows = preferences.nbRows;
+            this._displayElements();
+        };
+        this._onHideComplete = () => {
+            if (!this.interceptHide && this.actor) {
+                Grid_Main.layoutManager.removeChrome(this.actor);
+            }
+            Grid_Main.layoutManager["_chrome"].updateRegions();
+        };
+        this._onShowComplete = () => {
+            Grid_Main.layoutManager["_chrome"].updateRegions();
+        };
+        this._onResize = () => {
+            refreshGrids();
+            if (preferences.autoclose) {
+                this.emit('hide-tiling');
+            }
+        };
+        this._onMouseEnter = () => {
+            if (!this.isEntered) {
+                this.elementsDelegate.reset();
+                this.isEntered = true;
+            }
+            return false;
+        };
+        this._onMouseLeave = () => {
+            let [x, y, mask] = global.get_pointer();
+            if ((this.elementsDelegate && (x <= this.actor.x || x >= this.actor.x + this.actor.width)) || (y <= this.actor.y || y >= this.actor.y + this.tableHeight)) {
+                this.isEntered = false;
+                this.elementsDelegate.reset();
+                refreshGrids();
+            }
+            return false;
+        };
+        this._globalKeyPressEvent = (actor, event) => {
+            if (event.get_key_symbol() === Clutter.Escape) {
+                hideTiling();
+                return true;
+            }
+            return false;
+        };
+        this._onSettingsButton = () => {
+            this.elementsDelegate.reset();
+        };
+        this._bindKeyControls = () => {
+            Grid_Main.keybindingManager.addHotKey('gTile-close', 'Escape', toggleTiling);
+            Grid_Main.keybindingManager.addHotKey('gTile-tile1', 'space', this._keyTile);
+            Grid_Main.keybindingManager.addHotKey('gTile-tile2', 'Return', this._keyTile);
+            for (let index in KEYCONTROL) {
+                if (objHasKey(KEYCONTROL, index)) {
+                    let key = KEYCONTROL[index];
+                    let type = index;
+                    Grid_Main.keybindingManager.addHotKey(type, key, () => this._onKeyPressEvent(type, key));
+                }
+            }
+        };
+        this._removeKeyControls = () => {
+            this.rowKey = -1;
+            this.colKey = -1;
+            Grid_Main.keybindingManager.removeHotKey('gTile-close');
+            Grid_Main.keybindingManager.removeHotKey('gTile-tile1');
+            Grid_Main.keybindingManager.removeHotKey('gTile-tile2');
+            for (let type in KEYCONTROL) {
+                Grid_Main.keybindingManager.removeHotKey(type);
+            }
+        };
+        this._onKeyPressEvent = (type, key) => {
+            let modifier = type.indexOf('meta', type.length - 4) !== -1;
+            if (modifier && this.keyElement) {
+                if (!this.elementsDelegate.activated) {
+                    this.keyElement._onButtonPress();
+                }
+            }
+            else if (this.keyElement) {
+                this.elementsDelegate.reset();
+            }
+            switch (type) {
+                case 'gTile-k-right':
+                case 'gTile-k-right-meta':
+                    if (this.colKey === this.cols - 1) {
+                        this._keyTileSwitch();
+                    }
+                    this.colKey = Math.min(this.colKey + 1, this.cols - 1);
+                    this.rowKey = this.rowKey === -1 ? 0 : this.rowKey;
+                    break;
+                case 'gTile-k-left':
+                case 'gTile-k-left-meta':
+                    if (this.colKey === 0) {
+                        this._keyTileSwitch();
+                    }
+                    this.colKey = Math.max(0, this.colKey - 1);
+                    break;
+                case 'gTile-k-up':
+                case 'gTile-k-up-meta':
+                    this.rowKey = Math.max(0, this.rowKey - 1);
+                    break;
+                case 'gTile-k-down':
+                case 'gTile-k-down-meta':
+                    this.rowKey = Math.min(this.rowKey + 1, this.rows - 1);
+                    this.colKey = this.colKey === -1 ? 0 : this.colKey;
+                    break;
+            }
+            this.keyElement = this.elements[this.rowKey] ? this.elements[this.rowKey][this.colKey] : null;
+            if (this.keyElement)
+                this.keyElement._onHoverChanged();
+        };
+        this._keyTile = () => {
+            if (this.keyElement) {
+                this.keyElement._onButtonPress();
+                this.keyElement._onButtonPress();
+                this.colKey = -1;
+                this.rowKey = -1;
+            }
+        };
+        this._keyTileSwitch = () => {
+            let key = getMonitorKey(this.monitor);
+            let candidate = null;
+            for (let k in grids) {
+                if (k === key) {
+                    continue;
+                }
+                candidate = grids[k];
+            }
+            if (candidate) {
+                candidate._bindKeyControls();
+            }
+        };
+        this._destroy = () => {
+            for (let r in this.elements) {
+                for (let c in this.elements[r]) {
+                    this.elements[r][c]._destroy();
+                }
+            }
+            this.elementsDelegate._destroy();
+            this.topbar._destroy();
+            this._removeKeyControls();
+            this.monitor = null;
+            this.rows = null;
+            this.title = null;
+            this.cols = null;
+        };
+        this.tableWidth = 220;
+        this.tableHeight = 200;
+        this.borderwidth = 2;
+        this.bindFns = {};
+        this.rowKey = -1;
+        this.colKey = -1;
+        this.actor = new Grid_St.BoxLayout({
+            vertical: true,
+            style_class: 'grid-panel',
+            reactive: true,
+            can_focus: true,
+            track_hover: true
+        });
+        this.actor.connect('enter-event', this._onMouseEnter);
+        this.actor.connect('leave-event', this._onMouseLeave);
+        this.topbar = new TopBar(title);
+        this.bottombar = new Grid_St.Table({
+            homogeneous: true,
+            style_class: 'bottom-box',
+            can_focus: true,
+            track_hover: true,
+            reactive: true,
+            width: this.tableWidth
+        });
+        this.veryBottomBar = new Grid_St.Table({
+            homogeneous: true,
+            style_class: 'bottom-box',
+            can_focus: true,
+            track_hover: true,
+            reactive: true,
+            width: this.tableWidth
+        });
+        this._initGridSettingsButtons();
+        this.table = new Grid_St.Table({
+            homogeneous: true,
+            style_class: 'table',
+            can_focus: true,
+            track_hover: true,
+            reactive: true,
+            width: this.tableWidth,
+            height: this.tableHeight
+        });
+        this.actor.add(this.topbar.actor, { x_fill: true });
+        this.actor.add(this.table, { x_fill: false });
+        this.actor.add(this.bottombar, { x_fill: true });
+        this.actor.add(this.veryBottomBar, { x_fill: true });
+        this.monitor = monitor;
+        this.monitor_idx = monitor_idx;
+        this.rows = rows;
+        this.title = title;
+        this.cols = cols;
+        this.isEntered = false;
+        if (true) {
+            let nbTotalSettings = 4;
+            if (!toggleSettingListener) {
+                toggleSettingListener = new ToggleSettingsButtonListener();
+            }
+            let toggle = new ToggleSettingsButton('animation', SETTINGS_ANIMATION);
+            toggle.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
+            this.veryBottomBar.add(toggle.actor, { row: 0, col: 0, x_fill: false, y_fill: false });
+            toggleSettingListener.addActor(toggle);
+            toggle = new ToggleSettingsButton('auto-close', SETTINGS_AUTO_CLOSE);
+            toggle.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
+            this.veryBottomBar.add(toggle.actor, { row: 0, col: 1, x_fill: false, y_fill: false });
+            toggleSettingListener.addActor(toggle);
+            let action = new AutoTileMainAndList(this);
+            action.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
+            this.veryBottomBar.add(action.actor, { row: 0, col: 2, x_fill: false, y_fill: false });
+            action.connect('resize-done', this._onResize);
+            let actionTwo = new AutoTileTwoList(this);
+            actionTwo.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
+            this.veryBottomBar.add(actionTwo.actor, { row: 0, col: 3, x_fill: false, y_fill: false });
+            actionTwo.connect('resize-done', this._onResize);
+        }
+        this.x = 0;
+        this.y = 0;
+        this.interceptHide = false;
+        this._displayElements();
+        this.normalScaleY = this.actor.scale_y;
+        this.normalScaleX = this.actor.scale_x;
+    }
+    set_position(x, y) {
+        this.x = x;
+        this.y = y;
+        this.actor.set_position(x, y);
+    }
+    show() {
+        this.interceptHide = true;
+        this.elementsDelegate.reset();
+        let time = preferences.animation ? 0.3 : 0;
+        this.actor.raise_top();
+        Grid_Main.layoutManager.removeChrome(this.actor);
+        Grid_Main.layoutManager.addChrome(this.actor);
+        this.actor.scale_y = 0;
+        if (time > 0) {
+            Grid_Tweener.addTween(this.actor, {
+                time: time,
+                opacity: 255,
+                visible: true,
+                transition: 'easeOutQuad',
+                scale_y: this.normalScaleY,
+                onComplete: this._onShowComplete
+            });
+        }
+        else {
+            this.actor.opacity = 255;
+            this.actor.visible = true;
+            this.actor.scale_y = this.normalScaleY;
+        }
+        this.interceptHide = false;
+        this._bindKeyControls();
+    }
+    hide(immediate) {
+        this._removeKeyControls();
+        this.elementsDelegate.reset();
+        let time = preferences.animation && !immediate ? 0.3 : 0;
+        if (time > 0) {
+            Grid_Tweener.addTween(this.actor, {
+                time: time,
+                opacity: 0,
+                visible: false,
+                scale_y: 0,
+                transition: 'easeOutQuad',
+                onComplete: this._onHideComplete
+            });
+        }
+        else {
+            this.actor.opacity = 0;
+            this.actor.visible = false;
+            this.actor.scale_y = 0;
+        }
+    }
+};
+Grid = Grid_decorate([
+    addSignals
+], Grid);
+
+;
+
+;// CONCATENATED MODULE: ./src/3_8/extension.ts
+
+
 const GObject = imports.gi.GObject;
 const Cinnamon = imports.gi.Cinnamon;
 const extension_St = imports.gi.St;
 const Meta = imports.gi.Meta;
-const Clutter = imports.gi.Clutter;
+const extension_Clutter = imports.gi.Clutter;
 const extension_Signals = imports.signals;
 const extension_Main = imports.ui.main;
 const extension_Tweener = imports.ui.tweener;
@@ -629,7 +966,6 @@ let focusMetaWindowConnections = {};
 let focusMetaWindowPrivateConnections = {};
 let tracker;
 let gridSettingsButton = [];
-let toggleSettingListener;
 const preferences = {};
 let settings;
 const initSettings = () => {
@@ -974,332 +1310,6 @@ const getFocusApp = () => {
 const isPrimaryMonitor = (monitor) => {
     return extension_Main.layoutManager.primaryMonitor === monitor;
 };
-let Grid = class Grid {
-    constructor(monitor_idx, monitor, title, cols, rows) {
-        this.tableWidth = 220;
-        this.tableHeight = 200;
-        this.borderwidth = 2;
-        this.bindFns = {};
-        this.rowKey = -1;
-        this.colKey = -1;
-        this.isEntered = false;
-        this.interceptHide = false;
-        this._initGridSettingsButtons = () => {
-            this.bottombar.destroy_children();
-            let rowNum = 0;
-            let colNum = 0;
-            for (var index = 0; index < gridSettingsButton.length; index++) {
-                if (colNum >= 4) {
-                    colNum = 0;
-                    rowNum += 2;
-                }
-                let button = gridSettingsButton[index];
-                button = new GridSettingsButton(button.text, button.cols, button.rows);
-                this.bottombar.add(button.actor, { row: rowNum, col: colNum, x_fill: false, y_fill: false });
-                button.actor.connect('notify::hover', this._onSettingsButton);
-                colNum++;
-            }
-        };
-        this._displayElements = () => {
-            this.elements = [];
-            let width = this.tableWidth / this.cols - 2 * this.borderwidth;
-            let height = this.tableHeight / this.rows - 2 * this.borderwidth;
-            this.elementsDelegate = new GridElementDelegate();
-            this.elementsDelegate.connect('resize-done', this._onResize);
-            for (let r = 0; r < this.rows; r++) {
-                for (let c = 0; c < this.cols; c++) {
-                    if (c === 0) {
-                        this.elements[r] = [];
-                    }
-                    let element = new GridElement(this.monitor, width, height, c, r, this.elementsDelegate);
-                    this.elements[r][c] = element;
-                    this.table.add(element.actor, { row: r, col: c, x_fill: false, y_fill: false });
-                    element.show();
-                }
-            }
-        };
-        this.refresh = () => {
-            this.table.destroy_all_children();
-            this.cols = preferences.nbCols;
-            this.rows = preferences.nbRows;
-            this._displayElements();
-        };
-        this._onHideComplete = () => {
-            if (!this.interceptHide && this.actor) {
-                extension_Main.layoutManager.removeChrome(this.actor);
-            }
-            extension_Main.layoutManager["_chrome"].updateRegions();
-        };
-        this._onShowComplete = () => {
-            extension_Main.layoutManager["_chrome"].updateRegions();
-        };
-        this._onResize = () => {
-            refreshGrids();
-            if (preferences.autoclose) {
-                this.emit('hide-tiling');
-            }
-        };
-        this._onMouseEnter = () => {
-            if (!this.isEntered) {
-                this.elementsDelegate.reset();
-                this.isEntered = true;
-            }
-            return false;
-        };
-        this._onMouseLeave = () => {
-            let [x, y, mask] = global.get_pointer();
-            if ((this.elementsDelegate && (x <= this.actor.x || x >= this.actor.x + this.actor.width)) || (y <= this.actor.y || y >= this.actor.y + this.tableHeight)) {
-                this.isEntered = false;
-                this.elementsDelegate.reset();
-                refreshGrids();
-            }
-            return false;
-        };
-        this._globalKeyPressEvent = (actor, event) => {
-            if (event.get_key_symbol() === Clutter.Escape) {
-                hideTiling();
-                return true;
-            }
-            return false;
-        };
-        this._onSettingsButton = () => {
-            this.elementsDelegate.reset();
-        };
-        this._bindKeyControls = () => {
-            extension_Main.keybindingManager.addHotKey('gTile-close', 'Escape', toggleTiling);
-            extension_Main.keybindingManager.addHotKey('gTile-tile1', 'space', this._keyTile);
-            extension_Main.keybindingManager.addHotKey('gTile-tile2', 'Return', this._keyTile);
-            for (let index in KEYCONTROL) {
-                if (objHasKey(KEYCONTROL, index)) {
-                    let key = KEYCONTROL[index];
-                    let type = index;
-                    extension_Main.keybindingManager.addHotKey(type, key, () => this._onKeyPressEvent(type, key));
-                }
-            }
-        };
-        this._removeKeyControls = () => {
-            this.rowKey = -1;
-            this.colKey = -1;
-            extension_Main.keybindingManager.removeHotKey('gTile-close');
-            extension_Main.keybindingManager.removeHotKey('gTile-tile1');
-            extension_Main.keybindingManager.removeHotKey('gTile-tile2');
-            for (let type in KEYCONTROL) {
-                extension_Main.keybindingManager.removeHotKey(type);
-            }
-        };
-        this._onKeyPressEvent = (type, key) => {
-            let modifier = type.indexOf('meta', type.length - 4) !== -1;
-            if (modifier && this.keyElement) {
-                if (!this.elementsDelegate.activated) {
-                    this.keyElement._onButtonPress();
-                }
-            }
-            else if (this.keyElement) {
-                this.elementsDelegate.reset();
-            }
-            switch (type) {
-                case 'gTile-k-right':
-                case 'gTile-k-right-meta':
-                    if (this.colKey === this.cols - 1) {
-                        this._keyTileSwitch();
-                    }
-                    this.colKey = Math.min(this.colKey + 1, this.cols - 1);
-                    this.rowKey = this.rowKey === -1 ? 0 : this.rowKey;
-                    break;
-                case 'gTile-k-left':
-                case 'gTile-k-left-meta':
-                    if (this.colKey === 0) {
-                        this._keyTileSwitch();
-                    }
-                    this.colKey = Math.max(0, this.colKey - 1);
-                    break;
-                case 'gTile-k-up':
-                case 'gTile-k-up-meta':
-                    this.rowKey = Math.max(0, this.rowKey - 1);
-                    break;
-                case 'gTile-k-down':
-                case 'gTile-k-down-meta':
-                    this.rowKey = Math.min(this.rowKey + 1, this.rows - 1);
-                    this.colKey = this.colKey === -1 ? 0 : this.colKey;
-                    break;
-            }
-            this.keyElement = this.elements[this.rowKey] ? this.elements[this.rowKey][this.colKey] : null;
-            if (this.keyElement)
-                this.keyElement._onHoverChanged();
-        };
-        this._keyTile = () => {
-            if (this.keyElement) {
-                this.keyElement._onButtonPress();
-                this.keyElement._onButtonPress();
-                this.colKey = -1;
-                this.rowKey = -1;
-            }
-        };
-        this._keyTileSwitch = () => {
-            let key = getMonitorKey(this.monitor);
-            let candidate = null;
-            for (let k in grids) {
-                if (k === key) {
-                    continue;
-                }
-                candidate = grids[k];
-            }
-            if (candidate) {
-                candidate._bindKeyControls();
-            }
-        };
-        this._destroy = () => {
-            for (let r in this.elements) {
-                for (let c in this.elements[r]) {
-                    this.elements[r][c]._destroy();
-                }
-            }
-            this.elementsDelegate._destroy();
-            this.topbar._destroy();
-            this._removeKeyControls();
-            this.monitor = null;
-            this.rows = null;
-            this.title = null;
-            this.cols = null;
-        };
-        this.tableWidth = 220;
-        this.tableHeight = 200;
-        this.borderwidth = 2;
-        this.bindFns = {};
-        this.rowKey = -1;
-        this.colKey = -1;
-        this.actor = new extension_St.BoxLayout({
-            vertical: true,
-            style_class: 'grid-panel',
-            reactive: true,
-            can_focus: true,
-            track_hover: true
-        });
-        this.actor.connect('enter-event', this._onMouseEnter);
-        this.actor.connect('leave-event', this._onMouseLeave);
-        this.topbar = new TopBar(title);
-        this.bottombar = new extension_St.Table({
-            homogeneous: true,
-            style_class: 'bottom-box',
-            can_focus: true,
-            track_hover: true,
-            reactive: true,
-            width: this.tableWidth
-        });
-        this.veryBottomBar = new extension_St.Table({
-            homogeneous: true,
-            style_class: 'bottom-box',
-            can_focus: true,
-            track_hover: true,
-            reactive: true,
-            width: this.tableWidth
-        });
-        this._initGridSettingsButtons();
-        this.table = new extension_St.Table({
-            homogeneous: true,
-            style_class: 'table',
-            can_focus: true,
-            track_hover: true,
-            reactive: true,
-            width: this.tableWidth,
-            height: this.tableHeight
-        });
-        this.actor.add(this.topbar.actor, { x_fill: true });
-        this.actor.add(this.table, { x_fill: false });
-        this.actor.add(this.bottombar, { x_fill: true });
-        this.actor.add(this.veryBottomBar, { x_fill: true });
-        this.monitor = monitor;
-        this.monitor_idx = monitor_idx;
-        this.rows = rows;
-        this.title = title;
-        this.cols = cols;
-        this.isEntered = false;
-        if (true) {
-            let nbTotalSettings = 4;
-            if (!toggleSettingListener) {
-                toggleSettingListener = new ToggleSettingsButtonListener();
-            }
-            let toggle = new ToggleSettingsButton('animation', SETTINGS_ANIMATION);
-            toggle.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
-            this.veryBottomBar.add(toggle.actor, { row: 0, col: 0, x_fill: false, y_fill: false });
-            toggleSettingListener.addActor(toggle);
-            toggle = new ToggleSettingsButton('auto-close', SETTINGS_AUTO_CLOSE);
-            toggle.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
-            this.veryBottomBar.add(toggle.actor, { row: 0, col: 1, x_fill: false, y_fill: false });
-            toggleSettingListener.addActor(toggle);
-            let action = new AutoTileMainAndList(this);
-            action.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
-            this.veryBottomBar.add(action.actor, { row: 0, col: 2, x_fill: false, y_fill: false });
-            action.connect('resize-done', this._onResize);
-            let actionTwo = new AutoTileTwoList(this);
-            actionTwo.actor.width = this.tableWidth / nbTotalSettings - this.borderwidth * 2;
-            this.veryBottomBar.add(actionTwo.actor, { row: 0, col: 3, x_fill: false, y_fill: false });
-            actionTwo.connect('resize-done', this._onResize);
-        }
-        this.x = 0;
-        this.y = 0;
-        this.interceptHide = false;
-        this._displayElements();
-        this.normalScaleY = this.actor.scale_y;
-        this.normalScaleX = this.actor.scale_x;
-    }
-    set_position(x, y) {
-        this.x = x;
-        this.y = y;
-        this.actor.set_position(x, y);
-    }
-    show() {
-        this.interceptHide = true;
-        this.elementsDelegate.reset();
-        let time = preferences.animation ? 0.3 : 0;
-        this.actor.raise_top();
-        extension_Main.layoutManager.removeChrome(this.actor);
-        extension_Main.layoutManager.addChrome(this.actor);
-        this.actor.scale_y = 0;
-        if (time > 0) {
-            extension_Tweener.addTween(this.actor, {
-                time: time,
-                opacity: 255,
-                visible: true,
-                transition: 'easeOutQuad',
-                scale_y: this.normalScaleY,
-                onComplete: this._onShowComplete
-            });
-        }
-        else {
-            this.actor.opacity = 255;
-            this.actor.visible = true;
-            this.actor.scale_y = this.normalScaleY;
-        }
-        this.interceptHide = false;
-        this._bindKeyControls();
-    }
-    hide(immediate) {
-        this._removeKeyControls();
-        this.elementsDelegate.reset();
-        let time = preferences.animation && !immediate ? 0.3 : 0;
-        if (time > 0) {
-            extension_Tweener.addTween(this.actor, {
-                time: time,
-                opacity: 0,
-                visible: false,
-                scale_y: 0,
-                transition: 'easeOutQuad',
-                onComplete: this._onHideComplete
-            });
-        }
-        else {
-            this.actor.opacity = 0;
-            this.actor.visible = false;
-            this.actor.scale_y = 0;
-        }
-    }
-};
-Grid = extension_decorate([
-    addSignals
-], Grid);
-
-;
 
 gtile = __webpack_exports__;
 /******/ })()
