+++ gTile@shuairan/src/3_8/ui/AutoTileTwoList.ts
@@ -1,6 +1,7 @@
-import { focusMetaWindow, getNotFocusedWindowsOfMonitor, getUsableScreenArea, Grid, move_resize_window, reset_window } from "../extension";
+import { focusMetaWindow, getNotFocusedWindowsOfMonitor, getUsableScreenArea, move_resize_window, reset_window } from "../extension";
 import { addSignals } from "../utils";
 import { ActionButton } from "./ActionButton";
+import { Grid } from "./Grid";
 
 
 @addSignals
